ai-backend/requirements.txt
google-generativeai
pytesseract==0.3.13
pdf2image==1.17.0
python-dotenv==1.0.1
